Move common preview to folder
[MAILPOET-2743]
